Explicitly cast to uint8_t
Apple clang-1300.0.29.3 (ships with macOS 11.6) requires to use
`-Wno-shorten-64-to-32` and `-Wsnoign-conversion`, or explicitly cast to
smaller type or signed type.

`immediate` is `int64_t` and `opsize` is 4 bits field which makes mask
clearly no more than 5 bits => cast to `uint8_t` is safe here and just
makes clang happy and code clean.

`buf[]` is also `uint8_t` and cast happened before but implicitly.

+++ src/pci_ahci.c
@@ -220,8 +220,8 @@ static inline void lba_to_msf(uint8_t *buf, int lba)
 {
 	lba += 150;
 	buf[0] = (uint8_t) ((lba / 75) / 60);
-	buf[1] = (lba / 75) % 60;
-	buf[2] = lba % 75;
+	buf[1] = (uint8_t) (lba / 75) % 60;
+	buf[2] = (uint8_t) (lba % 75);
 }
 
 /*
